backend/pyproject.toml
[project]
name = "deer-flow"
version = "0.1.0"
description = "LangGraph-based AI agent system with sandbox execution capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deerflow-harness",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.1.0",
    "uvicorn[standard]>=0.34.0",
    "lark-oapi>=1.4.0",
    "slack-sdk>=3.33.0",
    "python-telegram-bot>=21.0",
    "langgraph-sdk>=0.1.51",
    "markdown-to-mrkdwn>=0.3.1",
    "pyyaml>=6.0.3",
    # 数据库
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",  # 数据库迁移（可选）
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动
    "aiosqlite>=0.19.0",  # SQLite 异步驱动
    # 安全相关
    "bleach>=6.0.0",  # HTML清理（用于XSS防护）
    "email-validator>=2.0.0",  # 邮箱验证
    # 可观测性相关
    "prometheus-client>=0.21.0",  # Prometheus指标
    "psutil>=5.9.0",  # 系统监控
    "langchain-core>=1.2.17",
    "python-dotenv>=1.2.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "starlette>=0.50.0",
    "stripe>=15.0.1",
    "alipay-sdk-python>=3.7.1018",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.14.11"]

[tool.uv.workspace]
members = ["packages/harness"]

[tool.uv.sources]
deerflow-harness = { workspace = true }
